requirements_lock.txt
blinker==1.8.2
bottle==0.12.25
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
proxy_tools==0.1.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-Quartz==10.3.1
pyobjc-framework-Security==10.3.1
pyobjc-framework-WebKit==10.3.1
pywebview==5.2
typing_extensions==4.12.2
Werkzeug==3.0.4
